requirements.txt
fastapi[standard]>=0.115.12
loguru>=0.7.3
opencv-python>=4.11.0.86
pydantic>=2.11.5
pydantic-settings>=2.9.1
python-dotenv>=1.1.0
pyyaml>=6.0.2
streamlit>=1.45.1
typer>=0.16.0
psutil
degirum>=0.17.2
degirum_tools>=0.18.0